+++ src/controller/config.ts
@@ -25,6 +25,8 @@ export class Config {
     readonly logLevel: LogLevel;
 
     readonly defaultEngine: TilingEngineType;
+    readonly insertInActive: boolean;
+
     readonly btreeSettings: object;
     readonly halfSettings: object;
 
@@ -45,9 +47,12 @@ export class Config {
         this.logLevel = rc("LogLevel", LogLevel.Warn);
 
         this.defaultEngine = rc("DefaultEngine", TilingEngineType.BTree);
+        this.insertInActive = rc("InsertInActive", false);
+
         this.btreeSettings = {
             swapInsertSide: rc("BTreeSwapInsertSide", false),
             rotateLayout: rc("BTreeRotateLayout", false),
+            depthFirst: rc("BTreeDepthFirst", false),
         };
         this.halfSettings = {
             swapInsertSide: rc("HalfSwapInsertSide", false),